+++ post-processor/vagrant/vmware.go
@@ -88,7 +88,7 @@ func (p *VMwareBoxPostProcessor) PostProcess(ui packer.Ui, artifact packer.Artif
 	}
 
 	// Create the metadata
-	metadata := map[string]string{"provider": "vmware"}
+	metadata := map[string]string{"provider": "vmware_desktop"}
 	if err := WriteMetadata(dir, metadata); err != nil {
 		return nil, err
 	}